+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.98.22", date: "2026-04-02" };
+  return { version: "8.98.23", date: "2026-04-02" };
 }